remove the atomic model in pt_expt. mv atomic model's output stat tests to dpmodel's folder.
