+++ deepmd/pt_expt/atomic_model/__init__.py
@@ -1,12 +1,1 @@
 # SPDX-License-Identifier: LGPL-3.0-or-later
-from .dp_atomic_model import (
-    DPAtomicModel,
-)
-from .energy_atomic_model import (
-    DPEnergyAtomicModel,
-)
-
-__all__ = [
-    "DPAtomicModel",
-    "DPEnergyAtomicModel",
-]
